+++ inst/source/R/zzz.R
@@ -1,0 +1,3 @@
+.onLoad <- function(libname, pkgname) { # Exclude Linting
+    run_r_tests_for_known_hosts()
+}