+++ inst/tinytest/test_fritools.R
@@ -8,7 +8,7 @@ f <- file.path(tempdir(), paste0("a", ".csv"))
 cars <- mtcars[1:2, TRUE]
 a <- write_csv(cars, file = f)
 expect_true(file.exists(f))
-expect_identical(get_path(a), f)
+expect_identical(strip_off_attributes(get_path(a)), f)
 
 #% get_path
 x <- 3
@@ -29,5 +29,5 @@ touch(tempfile)
 expect_error(set_path(x, tempfile()))
 # overwrite
 x <- set_path(x, tempfile, overwrite = TRUE)
-result <- get_path(x)
+result <- strip_off_attributes(get_path(x))
 expect_identical(result, tempfile)