+++ tests/testthat/test-csv.R
@@ -5,9 +5,10 @@ testthat::test_that("foo", {
     cars <- mtcars[1:2, TRUE]
     a <- write_csv(cars, file = f)
     testthat::expect_true(file.exists(f))
-    testthat::expect_identical(get_path(a), f)
+    testthat::expect_identical(strip_off_attributes(get_path(a)), f)
     p <- get_path(a)
     g <- paste0(f, "g")
-    testthat::expect_failure(testthat::expect_identical(p, g))
+    testthat::expect_failure(testthat::expect_identical(strip_off_attributes(p),
+                                                        g))
 }
 )